Personal Fields was None
Script had assumed the "personal_fields" attribute would always be a _dict. In one instance, it was discovered to be None. This correlated with the employee leaving the company, so likely this was a consequence of off-boarding the employee and disabling their account.

When personal fields is None, an error will not be thrown. The attributes will simply be missing from the object. To do: pre-define the fields so that in the case the personal_fields is None, the attributes will still be available and not cause issues if someone were to iterate through the user objects and reference data from personal _fields.